services/llm-python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mangum==0.19.0          # ASGI → Lambda adapter
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9  # multipart/form-data uploads

# Document parsing
pypdf==4.3.1
python-docx==1.1.2

# AWS
boto3==1.35.0
botocore==1.35.0

# Vector store (local dev)
qdrant-client==1.11.0

# Utilities
python-dotenv==1.0.1
